Add a  focusable prop to dropdown-item
Ref: #1437

Proposed changes:
The default behavior for dropdown item right now is to be focusable.
This prop will keep the default behavior, but can be disabled.

The login example in the doc is a good example.
In this case, we do not really expect the item to get focus.
Instead, we want the first input to be.