Fix datepicker for Safari (#2417)
* DatePicker fix for Safari

Update the defaultDateFormatter in the datepicker to handle dates far in the past on Safari (iOS and OS X). Refs #1957.

* Fix iOS native date picker

onChangeNativePicker gets a YYYY-MM-DD string from the native date control. Parse this manually for construction of the Date object from it, since iOS/Safari interprets the date as being UTC regardless of the presence of the time string:

> new Date("2020-04-11T00:00")
> Fri Apr 10 2020 19:00:00 GMT-0500 (CDT)
> new Date("2020-04-11")
> Fri Apr 10 2020 19:00:00 GMT-0500 (CDT)

Chrome interprets the first string format as being in local time.

* Add test cases for onChangeNativePicker

* Adjust testcase

* Adjust testcase

I should figure out how to run these tests locally..

* Adjust testcase

* Fix lint errors

* Change native date picker to use midnight