Fix null column in sorting data on sortable multiple table update data
When sortable multiple table was updated, the watcher of `data` data will call `sort` method with an argument of `currentSortColumn` data. But in sortable multiple table, `currentSortColumn` data is always an empty object. So, when the `sort` method call `sortMultiColumn` method, `sortMultiColumn` method will not find the column of `currentSortColumn` data (that get passed by `sort` method`), and make it thought that "this column is a new entry" and try to push into `sortMultipleDataLocal` data. Thus, an "undefined" column is now get sorting with others and cannot be remove - mess up with the number when rendered.

P.S. Apologies for my commit message.