[fix] Registered iconSize prop for input to pass onto icon
Currently there is a method called `iconSize` that is being passed to the `b-icon` component but its not being initialized anywhere or being accepted as a prop.

This commit allows the `input` component to receive iconSize as a prop with a default value of `is-large` so as to not change existing behavior.