Better support for wrapper around Autocomplete.vue
We are currently trying to make a wrapper component around `<b-autocomplete>` like so:

```js

// Our custom Autocomplete wrapper

<b-autocomplete
  v-model="query"
  :data="filteredTags"
  :icon="icon"
  :placeholder="placeholder"
  :field="field"
  @select="onInput"
  @typing="onTyping"
  >
      <template
        v-if="$scopedSlots.default"
        slot-scope="props"
      >
        <slot v-bind="props" />
    </template>
</b-autocomplete>
```

As you can see, we want to pass down the `default` slot when it's not empty. It's currently not possible to make it work with the current implementation. 

Instead of using the `hasDefaultSlot`, I propose the use of "default slots" to allow our use case. Was there a specific reason to use `hasDefaultSlot`?

I understand our use-case is very specific, let me know what you think! :)