Close tooltip when button clicked
If a click on the object the tooltip is encapsulating switches to a different screen, the tooltip can remain open. (and won't close until you return to the object, hover over it and hover-off) When the tooltip detects a click, if the tooltip is already open and the click is about to cause a switch away circumventing the default hover / blur close, then enforce a close first.