docs(CHANGELOG): prepare CHANGELOG for 0.1.3 (ntohq/buefy-next#224)
Prepares `CHANGELOG` for the release of v0.1.3.
- Adds missing changes in v0.1.3
- Adds missing changes in v0.1.2
- Removes the "unreleased" tag from v0.1.3

I decided to prefix "B" to component names, because user-facing
component names start with "b-". In addition, if users want to import
individual components, their names start with "B". So our communication
will be more consistent if we prefix "B" to component names.