docs: update CHANGELOG
I rewrite the CHANGELOG according to my preference, however we should
discuss and establish a minimum guideline for the CHANGELOG.

Categorizes the issue ntohq/buefy-next#19 as a "Breaking changes".

Removes "Fix" prefixes in the "Fixes" section because they look
redundant. Writes the issues instead of the solutions in the headlines.

Moves the issue ntohq/buefy-next#178 to a new category "Fixes (docs)"
because it does not matter to the core buefy-next library.

Adds the following fixes:
- ntohq/buefy-next#23
- ntohq/buefy-next#257
- ntohq/buefy-next#171 (docs)