TypeScript Rollout Tier 3 - config (#302) (#326)
* chore(lib): rename utils/config.js → .ts

Renames `utils/config.js` to `utils/config.ts` without correcting
TypeScript errors. This is to track changes before and after rewriting
it in TypeScript.

* feat(lib): rewrite utils/config in TS

Actually rewrites `utils/config.ts` in TypeScript:
Introduces new types:
- `BuefyConfig`: type for the entire Buefy config. I will refine the
  type of `BuefyConfig.customIconPacks` while rewriting
  `utils/icons.ts`
- `BuefyConfigOptions`: `BuefyConfig` with all the properties optional
- `ModalCancellableOption`
- `ModalScrollOption`
- `FieldLabelPosition`
- `NoticePosition`: possible values are defined in `NOTICE_POSITIONS`

Introduces a new type `VueClassAttribute` that represents the type for
values that can be bound to "class" attributes of DOM elements, since
Vue does not provide it. There might be better place for this type, but
defines it in `utils/config.ts` for now.

`index.ts` uses `BuefyConfigOptions` to represent the type of `options`
parameter of the `install` plugin method. It also exports types exported
from 'utils/config.ts'. Adds the type parameter to the `merge` call.

* test(lib): rewrite utils/config.spec in TS

- Rewrites `utils/config.spec.js` in TypeScript →
  `utils/config.spec.ts`. Imports necessary functions from `vitest`.