TypeScript Rollout Tier 4 - Provider / Injected / Tabbed mixins (#330)
* chore(lib): rename ProviderParent/InjectedChild mixins for TS

Replaces the extension of the following files, ".js" → ".ts":
- `src/utils/InjectedChildMixin.js`
- `src/utils/ProviderParentMixin.js`

This commit does not actually rewrite them in TypeScript, but maintains
their change logs accross renaming. Because massive changes that prevent
Git from recognizing renaming are anticipated.

* feat(lib): rewrite ProviderParent/InjectedChild mixins in TS

Rewrites the following mixins in TypeScript:
- `src/utils/ProviderParentMixin.ts` (`ProviderParentMixin`), which was
  just renamed at the last commit
- `src/utils/InjectedChildMixin.ts` (`InjectedChildMixin`), which was
  just renamed at the last commit
- `src/utils/TabbedMixin.js` (`TabbedMixin`)
- `src/utils/TabbedChildMixin.js` (`TabbedChildMixin`)

Introduces the following new files:
- `src/utils/ProvideInjectTypes.ts`: defines common types between
  `ProviderParentMixin` and `InjectedChildMixin`
- `src/utils/TabbedTypes.ts`: defines common types between `TabbedMixin`
  and `TabbedChildMixin`

Non trivial change in `ProviderParentMixin`:
- The default value of the `flags` parameter: 0 → `undefined`. Because
  the `flags` parameter is bound to the `FeatureFlags` type parameter
  and there is no guarantee that `FeatureFlags` accepts 0.

Non trivial changes in `InjectedChildMixin`:
- The default value of the `flags` parameter: 0 → `undefined`. Because
  the `flags` parameter is bound to the `FeatureFlags` type parameter
  and there is no guarantee that `FeatureFlags` accepts 0.
- Simplifies the type of the `dynamicIndex` data field as `number |
  undefined`, and its initial value becomes `null` → `undefined`

Wraps the entire definition of `TabbedMixin` and `TabbedChildMixin` with
a `defineComponent` call. `ProviderParentMixin` and `InjectedChildMixin`
are conditionally typed using the `DefineComponent` Vue utility type,
because their actual type depends on function parameters.  Both
`TabbedMixin` and `TabbedChildMixin` accept a type parameter so that
host components can specialize the child type and the parent type
respectively.

The following change is trivial but I leave the reason for it:
- Replaces JSDoc style comments with oridinary comments so that
  `@microsoft/api-extractor` picks them up as documentation