TypeScript Rollout Tier 5 - Collapse (#334)
* feat(lib): rewrite Collapse in TS

Rewrites `Collapse` in the `src/components/collapse` folder in
TypeScript:
- Introduces a new type `CollapsePosition` which represents the type of
  the `position` prop. Possible values are defined as a constant
  `COLLAPSE_POSITIONS`.

Replaces a JSDoc-style comment with an ordinary one so that
`@microsoft/api-extractor` won't pick it up for documentation.

* test(lib): rewrite Collapse.spec in TS

Rewrites the spec for `Collapse` in the `src/components/collapse` folder
in TypeScript.

All the changes are straightforward. Here is a tip for Jest → Vitest
migration:
- Import the spec building blocks from the `vitest` package

In the `__snapshots__` subfolder, replaces the spec snapshot produced by
Jest with the one by Vitest:
- `Collapse.spec.(js → ts).snap`

* chore(lib): bundle collapse in TS

`rollup.config.mjs` removes "collapse" from `JS_COMPONENTS`.

* feat(docs): rewrite collapse docs in TS

Rewrites the documentation for `Collapse` in the
`src/pages/components/collapse` in TypeScript.

All the changes are straightforward. Here is a tip for TypeScript
migration:
- Explicitly import and register components so that they are type
  checked. Note that no type-checking is performed for globally
  registered components.