TypeScript Rollout Tier 5 - Skeleton (#335)
* feat(lib): rewrite Skeleton in TS

Rewrites `Skeleton` in the `src/components/skeleton` folder in
TypeScript:
- The type definition of `BSkeleton` is awkward as it is forcefully
  casted as a Vue component via `unknown`. However, this is the only way
  I was able to come up with to type a functional component.
- Introduces a new interface `SkeletonProps` which represents the props
  type of `Skeleton`
- Introduces a new type `SkeletonPosition` which represents the type of
  the `position` prop. The possible values are given as a constant
  `SKELETON_POSITIONS`.

* test(lib): rewrite Skeleton.spec in TS

Rewrites the spec for `Skeleton` in the `src/components/skeleton` folder
in TypeScript.

All the changes are straightforward. Here is a tip for Jest → Vitest
migration:
- Import the spec building blocks from the `vitest` package

In the `__snapshots__` subfolder, replaces the spec snapshot produced by
Jest with Vitest:
- `Skeleton.spec.(js → ts).snap`

* chore(lib): bundle skeleton in TS

`rollup.config.mjs` removes "skeleton" from `JS_COMPONENTS`.

* feat(docs): rewrite skeleton docs in TS

Rewrites the documentation for `Skeleton` in the
`src/pages/components/skeleton` folder in TypeScript.

All the changes are straightforward. Here is a tip for TypeScript
migration:
- Explicitly import and register components so that they are type
  checked. Note that no type-checking is performed for globally
  registered components.