TypeScript Rollout Tier 5 - Breadcrumb (#344)
* feat(lib): rewrite breadcrumb in TS

Rewrites the breadcrumb components in the `src/components/breadcrumb`
folder in TypeScript.

All the changes are straightforward.

* test(lib): rewrite breadcrumb specs in TS

Rewrites the specs for the breadcrumb components in the
`src/components/breadcrumb` folder in TypeScript.

All the changes are straightforward. Here is a tip for Jest → Vitest
migration:
- Import the spec building blocks from the `vitest` package

In the `__snapshots__` subfolder, replaces the spec snapshots produced
by Jest with those by Vitest:
- `Breadcrumb.spec.(js → ts).snap`
- `BreadcrumbItem.spec.(js → ts).snap`

* chore(lib): bundle breadcrumb in TS

`rollup.config.mjs` removes "breadcrumb" from `JS_COMPONENTS`.

* feat(docs): rewrite breadcrumb docs in TS

Rewrites the documentation for the breadcrumb components in the
`src/pages/components/breadcrumb` in TypeScript.

All the changes are straightforward. Here is a tip for TypeScript
migration:
- Explicitly import and register components so that they are type
  checked. Note that no type-checking is performed for globally
  registered components.