TypeScript Rollout Tier 9 - Clockpicker (#376)
* feat(lib): rewrite clockpicker in TS

Rewrites the clockpicker components in the `src/components/clockpicker`
folder in TypeScript.

In `ClockpickerFace.vue`:
- Introduces a new interface `FaceNumber` which represents the item type
  of the `faceNumbers` prop

Common changes:
- Registers other Buefy components with "direct" names instead of the
  `name` fields so that they are type-checked

* test(lib): rewrite clockpicker spec in TS

Rewrites the specs for the clockpicker components in the
`src/components/clockpicker` in TypeScript.

All the changes are straightforward. Here are common changes:
- Imports the spec building blocks from the `vitest` package
- Replaces `jest` with `vi`

In the `__snapshots__` subfolder, replaces the spec snapshots produced
by Jest with those by Vitest:
- `Clockpicker.spec.(js → ts).snap`
- `ClockpickerFace.spec.(js → ts).snap`

* chore(lib): bundle clockpicker in TS

`rollup.config.mjs` removes "clockpicker" from `JS_COMPONENTS`.

* feat(docs): rewrite clockpicker docs in TS

Rewrites the documentation for the clockpicker components in the
`src/pages/components/clockpicker` in TypeScript. All the changes are
straightforward.

Here is a TypeScript migration tip:
- Explicitly import and register components so that they are type
  checked. No type-checking is performed for globally registered
  components.