TypeScript Rollout Tier 10 - Datetimepicker (#380)
* feat(lib): rewrite Datetimepicker in TS

Rewrites `Datetimepicker` in the `src/components/datetimepicker` folder
in TypeScript:
- `modelValue` allow `null`
- Replaces `isNaN(date)` with `isNaN(date.valueOf())` which should
  produce an equivalent result
- Introduces new types:
  - `DatetimeFormatter`: type of the `datetimeFormatter` prop
  - `DatetimeParser`: type of the `datetimeParser` prop
  - `DatetimeCreator`: type of the `datetimeCreator` prop

Here is a tip for TypeScript migration:
- Register Buefy components with "direct" names rather than the `name`
  fields so that they are type-checked

* test(lib): rewrites Datetimepicker.spec in TS

Rewrites the spec for `Datetimepicker` in the
`src/components/datetimepicker` folder in TypeScript.

All the changes are straightforward. Here are a tip for Jest → Vitest
migration:
- Import the spec building blocks from the `vitest` package

* chore(lib): bundle datetimepicker in TS

`rollup.config.mjs` removes "datetimepicker" from `JS_COMPONENTS`.

* feat(docs): rewrite datetimepicker docs in TS

Rewrites the documentation for `Datetimepicker` in the
`src/pages/components/datetimepicker` folder in TypeScript.

In `examples/ExEditable.vue`:
- Adds a missing field `enableSeconds` to `data`, which is `false`
  anyway

Other changes are straightforward. Here is a tip for TypeScript
migration:
- Explicitly import and register components so that they are type
  checked. Note that no type-checking is performed for globally
  registered components.