feat(docs): rewrite installation/Start.vue in TS (#401)
Rewrites `src/pages/installation/Start.vue` in TypeScript.

Fixes the issue that Vite complained about the missing `buefy` package,
which was not actually used. This was caused by the code snippets in
`Start.vue`, which Vite conufsed with the actual code and tried to
translate them. Moves these code snippets to a separate TypeScript file
`usage/start.ts`.

Here is a tip for TypeScript migration:
- Explicitly import and register Buefy components so that they are
  type-checked. Note that no type-checking is performed for globally
  registered components.