feat(docs): rewrite installation/Customization in TS (#400)
Rewrites `src/pages/installation/Customization.vue` in TypeScript.

Fixes the issue that Vite complained about the missing `buefy` package,
which was not actually used. This was caused by the code snippets in
`Customization.vue`, which Vite conufsed with the actual code and tried
to translate them. Moves these code snippets to a separate TypeScript
file `usage/customization.ts`.

Here is a tip for TypeScript migration:
- Explicitly import and register Buefy components so that they are
  type-checked. Note that no type-checking is performed for globally
  registered components.