Fix: Infinite Table updates? #191 (#436)
* test(lib): fix infinite updates in TableColumn spec

Fixes the issue that a test case for `TableColumn` caused infinite
updates. The test case specified an arrow function to the `td-attrs`
prop, and it caused the infinite updates. Replacing it with a component
instance method fixes the problem.

* chore(docs): add warning message to td-attrs of TableColumn

The description of the `td-attrs` prop of `TableColumn` shows a warning
message that tells that specifying an inline arrow function leads to
infinite updates.

* docs: add a record regarding infinite Table updates to CHANGELOG

* fix(docs): type errors in a fresh clone

Fixes the issue that a bunch of type errors similar to the following
were produced in a fresh clone:

    Error: src/App.vue(33,22): error TS2339: Property '$buefy' does not exist on type 'CreateComponentPublicInstance<{}, {}, { clipboard: ClipboardJS | null; }, {}, { setupClipboardControls(): void; }, ComponentOptionsMixin, ComponentOptionsMixin, ... 12 more ..., {}>'.

While I made a change to utilize the Project Reference feature of
TypeScript in the previous commit
8b513e7a1966bad1f21edee7f00b62b08c5f33d4, it required to build the
`buefy-next` package prior to the `docs` project. This is inconvenient
for the development purpose. I decided to give up the Project Reference
feature and directly include source files in the `../buefy-next/src`
folder. As the `../buefy-next/src` folder is included in `rootDirs` this
time, it is no longer illegal to import files from there.