Fix: [Table] Sort multiple example does not work #434 (#437)
* fix(lib): multi column sorting crashed

Fixes the bug that multi column sorting crashed. It crashed at the
`getSortingObjectOfColumn` method because the `column` parameter which
was supposed to be non-`null` was `null`. The `column` parameter became
`null` when the `sortMultipleSelect` data field was `null` (initial
value) and given to the `columnIsDesc` method which invokes
`getSortingObjectOfColumn`. The initial value of the
`sortMultipleSelect` data field used to be an empty string ('') but was
replaced with `null` during the TypeScript migration, because an empty
string did not make sense at all and caused a type error. `columnIsDesc`
now makes sure that the `column` argument is not `null` before passing
it to `getSortingObjectOfColumn`.

* docs: add a record regarding Table multi sorting to CHANGELOG