Publish Buefy v1.0.0 (#4077)
* chore: rename buefy-next back to buefy

* chore: update package.json for v1.0.0

* chore: rename library folder to buefy

* chore: rename buefy-next with buefy

* chore: update package-lock.json

* chore: fix lint issues

* chore: update unit test snapshots

* chore: update readme

* chore: update changelog

* chore: unscope NPM publish pipeline

* chore: update buefy package readme

* fix: clean up NPM pipeline indentation

* chore: remove mention of v3 subdomain

* chore(gha): include scope `@buefy` in dev package name

Due to the GitHub npm registry requirement, the `publish_dev_to_gpr`
workflow has to include the scope, i.e., organization name, in the
package name to publish it to the GitHub npm registry. Runs
`npm install` instead of `npm ci`, because `npm ci` won't work due to
the package name change. Also replaces the namespace specified to the
build command, and publish command.

* chore(gha): revert tab size: 4 → 2 spaces

* docs: update README

Updates the section for installation of a developer release.

---------

Co-authored-by: Kikuo Emoto <kemoto@codemonger.io>