fix(lib): Tooltip was shown at a wrong place if always
Fixes an issue that `Tooltip` was shown at a wrong place until a user
interacted with it if the `always` prop is `true`. Fixes it by
determining `dynamicPosition` on mounted if the `always` prop is `true`.

Background:
If the `always` prop is `true`, the `Tooltip` is visible on mounted with
the `dynamicPosition` data field `undefined`. The `dynamicPosition` data
field was not calculated until the `open` method or the
`updateAppendToBody` method was called; e.g., until a user interacted
with a component wrapped in `Toolbox`.