fix(lib): Table did not emit mouseenter/leave
Fixes the issue that `Table` did not emit "mouseenter", and "mouseleave"
events by
- removing "mouseenter" and "mouseleave" from the `emits` listing,
  because they are "specially" handled
- fixing the listener name used by the `emitEventForRow` function

"mouseenter", and "mouseleave" events are specially handled by the
`emitEventForRow` function. The problem was the `emitEventForRow`
function tried to get an event listener from `$attrs` with a wrong name.
For instance, the old code specified "onmouseenter" while "onMouseenter"
was expected.

The `emitEventForRow` function was introduced by the PR #3236. But the
original idea of the special treatment of events dates back to the PR
whether the justification of the PR is still relevant to Vue 3 but
retain it for now.